Quickfix: Restrict tinyrpc to version 0.9.4
tinyrpc 1.0.0 breaks python 2.7 support and only supports python3. 0.9.4
is the last version with python 2.7 support. Idealy this should be fixed
in the dependency list of ryu, but until that happens the problem can be
worked around by manually installing version 0.9.4.

Without this fix newly created virtual machines won't be able to run the
sdn controller